requirements.txt
# Core
Flask==2.0.1
Werkzeug==2.0.1
gunicorn==20.1.0
redis==4.0.2
SQLAlchemy==1.4.23

# Security
Flask-JWT-Extended==4.3.1
bcrypt==3.2.0
Flask-Limiter==2.4.0

# Frontend
Jinja2==3.0.1

# Testing
pytest==6.2.5
pytest-cov==2.12.1
locust==2.8.3

# Monitoring
prometheus-client==0.11.0

Flask-SQLAlchemy==2.5.1
Flask-Migrate==3.1.0
Flask-RESTx==0.5.1
marshmallow==3.14.0
Pillow==8.4.0
boto3==1.20.24
redis==4.0.2
apscheduler==3.8.1